Fixed an issue with renders width and height. Updated collision.

+++ ThunderEngine/src/com/MightyChubz/core/mathf/Vector2f.java
@@ -9,7 +9,7 @@ public class Vector2f {
 
     public float x, y;
 
-    public Vector2f(int x, int y) {
+    public Vector2f(float x, float y) {
         this.x = x;
         this.y = y;
     }
